+++ test/MOI_wrapper_direct.jl
@@ -58,4 +58,7 @@ end
     MOIT.int1test(OPTIMIZER, CONFIG)
     MOIT.int2test(OPTIMIZER, CONFIG)
     MOIT.int3test(OPTIMIZER, CONFIG)
+    MOIT.indicator1_test(OPTIMIZER, CONFIG)
+    MOIT.indicator2_test(OPTIMIZER, CONFIG)
+    # MOIT.indicator3_test(OPTIMIZER, CONFIG) # no support for ACTIVATE_ON_ZERO
 end